+++ frontend/src/containers/Settings/Settings.module.css
@@ -1,0 +1,38 @@
+.Settings {
+    height: 2rem;
+    width: 2rem;
+    z-index: 10;
+}
+
+.Settings_btn {
+    background: url('../../assets/settings.svg') center/2rem no-repeat;      
+    height: 2rem;
+    width: 2rem;
+    border: none;
+    cursor: pointer;
+}
+
+.Settings_menu {
+    position: absolute;
+    bottom: 4rem;
+    left: 0;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    z-index: 1;
+    background: white;
+    border: none;    
+    width: 100%;   
+}
+
+.Settings_menu_item {
+    background: slategray;
+    border: none;
+    color: white;
+    text-decoration: none;
+    font-size: 1.5rem;
+    flex: 1;
+    text-align: center;
+    padding: 0.5rem;
+    cursor: pointer;
+}
