--- frontend/src/containers/Settings/SettingsMenu/SettingsMenu.js
@@ -1,12 +1,0 @@
-import React from 'react'
-
-const SettingsMenu = props => {
-    return (
-        <div className="settings-menu">
-            <button className="about-btn">About</button>
-            <button className="signout-btn">Sign Out</button>
-        </div>
-    )
-}
-
-export default SettingsMenu
